Varying group/event data, follow String for uid, create UserList class
