+++ lib/model/user_card_search.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:connectuni/model/user.dart';
-import 'users.dart';
+import 'userList.dart';
 
 class UserCardSearch extends StatelessWidget {
   const UserCardSearch({Key? key, required this.name}) : super(key: key);
