+++ lib/screens/group_chat_screen.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import '../model/group.dart';
 import '../model/message_model.dart';
 import '../model/user.dart';
-import '../model/users.dart';
+import '../model/userList.dart';
 import 'groupinfo.dart';
 
 class GroupChatScreen extends StatefulWidget {
