+++ lib/screens/search_people_screen.dart
@@ -1,4 +1,4 @@
-import 'package:connectuni/model/users.dart';
+import 'package:connectuni/model/userList.dart';
 import 'package:connectuni/model/user_card_search.dart';
 import 'package:flutter/material.dart';
 import 'package:multi_select_flutter/multi_select_flutter.dart';
